+++ src/Command/ProcessQueue.php
@@ -6,7 +6,7 @@ namespace Dragonmantank\Sched\Command;
 
 use DI\Annotation\Inject;
 use Dragonmantank\Sched\LoggingTrait;
-use Pheanstalk\Pheanstalk;
+use Dragonmantank\Sched\Queue\QueueService;
 use Psr\Container\ContainerInterface;
 use Psr\Log\LoggerInterface;
 use Psr\Log\LogLevel;
@@ -37,7 +37,7 @@ class ProcessQueue extends Command
      */
     public function __construct(
         protected array $config,
-        protected Pheanstalk $pheanstalk,
+        protected QueueService $queueService,
         protected ContainerInterface $container,
         protected ?LoggerInterface $logger,
     ) {
@@ -58,18 +58,17 @@ class ProcessQueue extends Command
         $queueName = $input->getArgument('queueName');
         $numberOfJobs = (int) $input->getOption('number');
 
-        $this->pheanstalk->watch($queueName);
         for ($i = 0; $i <= $numberOfJobs; $i++) {
             $this->log($output, LogLevel::DEBUG, 'Waiting for Job ' . $i . ' in ' . $queueName);
-            $stats = $this->pheanstalk->statsTube($queueName);
-            if ($stats['current-jobs-ready'] < 1) {
+            $stats = $this->queueService->getStats($queueName);
+            if ($stats[$queueName]['current-jobs-ready'] < 1) {
                 exit(0);
             }
 
-            $job = $this->pheanstalk->reserve();
+            $message = $this->queueService->receiveMessage($queueName);
             try {
-                $this->log($output, LogLevel::DEBUG, 'Received job ' . $job->getId() . ' in ' . $queueName);
-                $payload = json_decode($job->getData(), true);
+                $this->log($output, LogLevel::DEBUG, 'Received job ' . $message->id . ' in ' . $queueName);
+                $payload = json_decode($message->payload, true);
                 $worker = $this->config['queues'][$queueName]['worker'];
 
                 if (is_string($worker)) {
@@ -84,24 +83,32 @@ class ProcessQueue extends Command
                 $exitCode = $worker($payload);
 
                 if ($exitCode === 0) {
-                    $this->log($output, LogLevel::DEBUG, 'Finished, deleting job ' . $job->getId() . ' from ' . $queueName);
-                    $this->pheanstalk->delete($job);
+                    $this->log($output, LogLevel::DEBUG, 'Finished, deleting job ' . $message->id . ' from ' . $queueName);
+                    $this->queueService->deleteMessage($queueName, $message);
                 } else {
                     $this->log($output, LogLevel::ERROR, 'Worker returned ' . $exitCode . ', rescheduling job for 60 seconds');
-                    $this->pheanstalk->release(job: $job, delay: 60);
+                    $stats = $this->queueService->getMessageStats($queueName, $message);
+                    $this->log($output, LogLevel::DEBUG, 'Job ' . $message->id . ' reserved ' . $stats['reserves'] . ' times');
+
+                    if ($stats['reserves'] > 3) {
+                        $this->log($output, LogLevel::DEBUG, 'Job ' . $message->id . ' buried due to bad worker results');
+                        $this->queueService->buryMessage($queueName, $message);
+                    } else {
+                        $this->queueService->releaseMessage(queueName: $queueName, message: $message, delay: 60);
+                    }
                 }
             } catch (\Exception $e) {
-                $this->log($output, LogLevel::DEBUG, 'Received error, releasing job ' . $job->getId() . ' from ' . $queueName);
+                $this->log($output, LogLevel::DEBUG, 'Received error, releasing job ' . $message->id . ' from ' . $queueName);
                 $this->log($output, LogLevel::ERROR, $e->getMessage());
-                $stats = $this->pheanstalk->statsJob($job);
-                $this->log($output, LogLevel::DEBUG, 'Job ' . $job->getId() . ' reserved ' . $stats['reserves'] . ' times');
+                $stats = $this->queueService->getMessageStats($queueName, $message);
+                $this->log($output, LogLevel::DEBUG, 'Job ' . $message->id . ' reserved ' . $stats['reserves'] . ' times');
 
                 if ($stats['reserves'] > 3) {
-                    $this->log($output, LogLevel::DEBUG, 'Job ' . $job->getId() . ' buried');
-                    $this->pheanstalk->bury($job);
+                    $this->log($output, LogLevel::DEBUG, 'Job ' . $message->id . ' buried');
+                    $this->queueService->buryMessage($queueName, $message);
                 }
 
-                $this->pheanstalk->release(job: $job, delay: 60);
+                $this->queueService->releaseMessage($queueName, $message, delay: 60);
             }
         }