+++ src/Exception/NoMessageInQueueException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dragonmantank\Sched\Exception;
+
+class NoMessageInQueueException extends \RuntimeException
+{
+}